Avoid reading the magnetic field from file if we load the geometry from the Database
The magnetic field is read from the compressed text file and stored in
memory in all cases but never used if the geometry is to be loaded from
the Database.

This commit fixes this issues by implementing the `createFromDB()`
member which does nothing. This overrides the base class member which
raises a `NotImplemented` exception and was thus falling back to the old
xml/text file procedure.