Fix merging of BackgroundInfo (BII-3049)
The merging of different BackgroundInfo objects was not reliable: If the
tag and fileType were identical but the filenames are different it would
add a new background but continue looking at all remaining backgrounds

* if we have [file1, file2] and merging this with [file2] we end up with
  [file1, file2, file2]
* this can in extreme cases lead to endless loops somehow because we
  even check the newly added backgrounds if we want to add the
  background ... again.

This commit changes the logic: It goes over all existing backgrounds
once. If a match is found, add to that one and break. If not we add it
to the list after the loop.