Fix memory leak in path.for_each (BII-2893)
It seems that we lost one instance of the looped-over type per event:
The objectEntry.object was already pointing to a valid instance but we
just overwrote that pointer and replaced it with a `new TObject`.

This commit just remembers the original value and resets it after the
`for_each`. This should also be slightly faster as with mutliple
`for_each` per event or if there are no objects to loop over but not
significantly.