Bugfix: "weight" in `CurlerCloneTruthVarSet` inverted, added `not`
Also removed the debugging code that was introduced to find this.

Bug explanation: The inverted weight selected fake tracks, so I looked
for loop numbers in fakes, which sometimes have not matched hits,
which resulted in the lookup of getFirstNLoops returning
c_InvalidIndex (=INT_MIN).

Note to self: When debugging, try not to think too much where the
error might be, just use debugging tools, log outputs and consider
unit tests.