Use checkPayloads call to check for existing payloads in the Database
This allows to check for the existence of payloads across global tags so
the script will not upload anything which already exists. Useful for
download/edit/upload cycle.

Also, allow for one payload to have multiple iovs when uploading,
something like

```
dbstore/VXDServiceGeometryPar 3 0,0,0,-1
dbstore/VXDServiceGeometryPar 3 1,0,5,-1
dbstore/VXDServiceGeometryPar 3 6,0,-1,-1
```

This kind of worked before but if the payload wasn't present in the
database it would have tried to create three independent payloads. Now
it should work in all cases.