Add authentication and a common user to b2conditionsdb
Yes, a hardcoded password in the python script is not a very secure
thing. This is just to make sure our server is not globally writable
without any password and will be replaced by proper authorization.

Optionally we could require a different username/password for management
style rest requests (delete everything in GT, mark public/invalid) which
could be supplied with the `--http-user`, `--http-password` by Marko.