Add "AVAILABLE_PACKAGES" to SCons environment
and disable `dqm` package unless `daq` is included in the list of
available packages.

This check in dqm can be moved to sub directories to be more fine
grained and only disable code which actually depends on the daq package
and not the whole package but at least this should get rid of the
problem that `newrel foo; setuprel; scons` does not work at the moment.