Changes needed for ConditionsDB transition
This commit changes a few things in the conditions db backend. Some
thing could be done more elegant but this is the least invasive version
due to the limited amount of time:

* Allow users to set a list of database servers to try in order. This
  can either be done with `basf2.set_central_serverlist()` which takes a
  python list of server urls or by setting the environment variable
  `BELLE2_CONDB_SERVERLIST` to all urls to try in turn, separated by whitespace.
* Change default to first try BNL and fall back to PNNL:
  `["http://blcdb.sdcc.bnl.gov/b2s/rest/", "http://belle2db.hep.pnnl.gov/b2s/rest/"]`
* Enable transparent redirect following in curl with a limit of 10
  redirects. This should allow to have BNL as first choice if redirect
  is setup on that server until switchover happens and just start
  serving transparently.

I still need to do more detailed testing but this is all that should be
necessary. We can still decide if we keep PNNL as primary server and
switch by (1) "turning PNNL off" or by (2) "turning BNL on".

1. will cause warnings as soon as PNNL is off but will then fall back to
   BNL without problems. Warnings could be avoided if PNNL issues
   redirects to BNL in the beginning
2. will cause warnings until BNL starts serving unless redirect is setup
   until we switch.

So the second option with redirects  would be preferred from my side as
in this case users should not notice the switch at all. In this case
the server list will never actually be used.

Additional changes in this commit as they seemed a good idea to also
add here:

* Disable SSL Verification for now: This should allow to switch to https
  even some grid sites would not have a proper CA bundle installed at
  the default location where curl is looking for it.
* Add a `User-Agent:` string to identify different versions to spot
  which release causes problems if any. This is not very detailed yet