Add additional output track parameter for TrackCreatorSingleSegments.
This can be identical to the inputTrack parameter (same behavior as
before), but allows to push the SingleSegmentTracks into a different
CDCTrackVector, if required.

The TrackCreatorSingleSegments had one input tracks parameter, which
was used to decide which segments are already used. When it decided to
promote a segment to a track, it was pushed back to the same track
vector.

For my studies on CDC track finders, I need to differentiate between
tracks found by different track finders, which I need the possibility
to push found SingleSegmentTracks to a specific CDCTrackVector. All
other track finder modules have the possibility to specify an output
track vector, which is why I added it here.

The downside is, that in most cases the input and output will be
identical, so this leads to more verbose code.