Try to fix constness of DBObjPtr
In principle contents of DBObjPtr should be non-modifyable. We already
have this in python but I just saw this is not enforced in C++.

This commit changes the defaul DBObjPtr and DBArray interfaces to return
const objects. However some packages do complicated things with DBObjPtr
and don't follow const correctness so I cannot just enable it everywhere

Instead I added a "DISABLE_CONST_DBOBJECTS" define which allows to keep
the old interface and enabled this for all packages which have offending
usage of db object pointers I could not quickly fix.

If this define is enabled in the SConscript every use of DBObjPtr data
will create a warning so unless the define is removed the warnings will
appear.

Once we are free of warnings I will remove the #ifdef in DBObjPtr.

Most changes in all packages are only fixes to const correctness but in
arich I added some const_casts in the ARICHDatabaseImporter. This code
should probably restructured.

In reconstruction I changed the getters to return const references to
thethe PDF instead of copying the pdf on return. The uses of the getters
should be checked to maybe accept the value by reference to avoid
copies.