Fix empty sphinx output (BII-3572)
b3code-module exits the program if it is imported. Since sphinx will
import it to get the program parameters for the documentation this will
also exit sphinx.

This just fixes the side-effect guard and then sphinx will run fine
again. Not sure yet how we can guard against this correctly to make sure
something like this doesn't happen in the future but I should start
running sphinx on bamboo and making sure it runs succesfully by
subscribing to the build-finished event in sphinx and creating a stamp
file or something similar.