Fix exessive memory usage by DB backend
It seems there are way more short lived, temporary DBObjPtr objects
around than I expected. Each of those temporaries would register a
callback and we would only clean up this list of callbacks when the
payload changed.

This commit changes this to a register/release system where the
DBAccessors register and unregister themselves with the DBStore. This
means that cleanup will be done immediately and callback execution does
not need to check validity anymore.

User registered callbacks are now owned and kept in the DBAccessor
objects so the lifetime is still bound to the DBAccessor but we don't
have to pass around shared_ptr anymore.