Reconnect EventMetaData for each run update (BII-3608)
Otherwise reset of DBStore and DataStore might be out of sync

* DBStore uses DataStore to determine event number
* DBStore checks the event meta data everytime a DBObjPtr is created. On
  doing so the EventMetaData StorObjPtr checks the DataStore.
  This might be before processing even starts.
* If there's already an existing EventMetaData from a previous
 `process()` call the StoreObjPtr is connected
* when starting processing the DataStore is reset and if the StoreObjPtr
  was connected it becomes invalid (BII-1262)

The contents of the DataStore are accessed after processing, for example
by the execution statistics so we cannot just move the reset to a
different place. But we can reconnect on each DBStore::update() which
happens once per run so the overhead should be small until (BII-1262) is
fixed.