Optimize b2conditionsdb-extract
* get rid of the `eval()`s, use `getattr` instead to obtain the members
  from the specific namespaces
* Don't use global variables, put everyting in the classes
* Clean up a bit

Still need to add/expand sphinx documentation.

There was a problem with assigning the new object with

```python
self.obj = self.dbobj.obj()
```

as this will change the pointer, not copy the contents. The better
solution would be to correctly update the branch address before filling
but I tried this and it segfaulted so for now this uses `obj.__assign__`
which calls the assignment operator on the objects.