Forward iteration for PyStoreObj and PyDBObj
If the managed object is iterable it should be transparent for the user
to iterate over it. Fix this by forwarding the iterator if it exists. If
not a type error is raised that the object is not iterable which would
be the same as iterating over a non-managed object.