Fix possible memory runaway in RelationIndex
For fast access to Relations we create a bidirectional lookup index on
demand, the RelationIndex. This index is rebuild every time it is
accessed if the underlying relations array has been modified.
That means we didn't clean the on each new event as they were rebuilt on
next access automatically.

Now as an optimization the RelationsObject::addRelation checks if an
index exists (which means someone was actually using the relation) and
if so it will add the relation directly to the index as well and not
mark the relation array as modified. This saves rebuilding of the index
for every small change.

Now if the relation is actually used then this is still fine: In the
next event the relations array is initialized as modified so the index
will be rebuilt on first access.

However, if for example the occupancy in the SVD is large then the
Spacepoints will be created with relations but the VXDTF will not run
due to the high occupancy. Also there are probably no CDC tracks so I
guess nobody is accessing the relations and we continuously just add
elements to the index ...

This patch fixes this problem by clearing all caches for each event.
Might have a slight negative impact on speed but basically it just moves
the clearing to be done earlier so I expect the effect to be small if at
all.