Fix problem with dangling DataStore accessors in tests and multiple process() calls
The last commit modified the clear() function to just clear the contents
of the RelationIndexContainers but keep the objects to not incur
overhead for the creation of these on each event.

However that caused a few issues when running tests or multiple calls to
`process()` as in that case the DataStore is reset completely in between
so we really have to clear the RelationIndexContainers completelty as
well.

Fixed that now.