Create the EventKinematics module and data object.
This commit is part of the event shape project.

In release 2, the EventShape module and the corresponding data object included mostly quantities that are not traditionally related to the event shape,
but to the event kinematics (total missing momentum, total visible energy...). The only actual event shape variable was the total thrust of the event.

With this commit I have created an EventKinematics module, that is basically a copy of the former EventShape module without the thrust calculation, so that now
all the actual "event shape" observable can be regrouped in the EventShape module, and all the "event kinematics" variables can be regrouped in the EventKinematics module

In the next commit, I will remove all the kinematic variables from the EventShape module, finalizing the complete separation between the EventShape and the EventKinematics.