Make LTO linking parallel to avoid sequential bottlenecks on big libs
The advantage of doing this is that small builds (e.g. those that occurs
after changing only a source file or two) are much faster, since the
overhead of re-doing LTO is spread across CPU cores. This speeds up the
development cycle very advantageously.

The drawback, however, is that this will cause some system
oversubscription if there _are_ operations which can be carried out in
parallel with the link. As a result, CPUs will be used slightly less
efficiently, but more importantly more RAM will be used and disk I/O
patterns will be less favorable.

To resolve this cleanly, GCC's LTO plugin and SCons would need to
communicate so that one does not spawn extra tasks if the CPU is already
fully saturated. There is support for this when using GNU Make (see
-flto=jobserver), but sadly SCons reinvents Make so this optimization
cannot be applied here.

A lesser alternative would be to make the number of LTO jobs
configurable by the user of SCons, e.g. via an env variable.