Make logging work in notebooks (BII-4059)
To make jupyter be able to see our log messages we have to send them to
pythons `sys.stdout` instead of directly printing them to the output. We
don't want to do this all the time since it might affect performance
but should be fine in case we run in a notebook.

So the LogConnectionFileDescriptor has been renamed to
LogConnectionConsole which takes a file descriptor for output but, in
case it has been enabled via `LogConnectionConsole::setEnablePythonLogging(true)`
or `basf2.logging.enable_python_logging = True`m it will send the output
to python instead of directly to the file descriptor.

Now with basf2 being a package we can check if we run in a notebook on
import and activate the feature automatically in a notebook.