Use json logging in jupyter processing
So far log messages were identified by the line starting with [INFO] (or
any other log level) which is not really robust. Now we use json
formatted log objects and can parse them correctly: each log entry is on
one line and if it is a json object then use that, otherwise show line
as is.