Add new FormulaParser using the shunting-yard algorithm.
This is adds a generic new class to parse strings of the form "a + b *
(c / d)" taking operator precedence into account. The operators
available are `+`,`-`,`*`,`/`,`^` and `**` as alias for `^`. The
operands can be either floating point numbers or string identifieres
with an optional argument list in brackets `()`.

The formula is parsed and converted in reverse polish notation with the
shunting-yard algorithm and then converted to the desired output object

The output can be configured using a `VariableCreator` struct. Currently
there is such a struct to convert the formula to a string representation
but the same could be easily implemented to convert to VariableManager
variables instead.

This would make it possible to write "formula((-3+5) * E ^ 2 / 3e-12)" and
this would be converted into the equivalent

```
  ((2.00 * pow('E',2)) / 3.00e-12))
```

Consecutive float operands are eagerly calculated on parsing so "1 + 5"
returns "6".

What this needs is

1. A **LOT** more tests in `framework/testing/formula_parser.cc`
2. A VariableManager interface
3. replace the old formula() code

(maybe become the default for VariableManager.addAlias(name, formula))