Include corrections number of empty_s hit gaps
`empty_s_size = size - 1`, because the number of hit gaps is one lower than the number of hits. This
changes the calculations of the variance and mean. It also means, that the empty_s variables are
only calculatable for size > 1 and the empty_s_variance is only calculatable for size > 2.

Also removed some values from an else clause, because the same effect can be achieved by simply
changing the initial values.