Change the way we strip debug symbols
SCons calculates a signature of the binary/libary after building so when
we modify it again in a later build step it will get confused and might
rebuild the libraries.

The official recommendation is to use `AddPostAction` which executes a
given action on the file after it is build but before the signature is
created.

To make sure that the directory exists and that the debug files are not
auto cleaned we add them as a `SideEffect` and we also add them to the
the list of files to be cleaned (although I don't think many people will
do scons -c).