Fix cppcheck warning about memory leak
cppcheck warns because it cannot guarantee that the object is freed ad
the end. While delete is called in `terminate()` it's not called in the
destructor which is the only place guaranteed to be called. To get rid
of this we can just move the chain in a `std::unique_ptr` which behaves
almost like a pointer except it takes ownership and deletes the object
when destructed.