Bugfix: toFinite now checks if var is finite as float
The function toFinite is used in the context of VarSets, to set infinite/nan values to 0. The
variable calculations are performed with doubles (for precision), but then stored in floats. Since
the toFinite function was defined for floats, it could happen, that a variable passes the isfinite
test for doubles, but is then converted to inf when it is cast to float. This has been fixed by
changing all arguments and the return value to float.

An alternative way to fix this would be to put a cast to float into the `std::isfinite` check, but
otherwise leave the arguments and return value as double, to prevent an eventual loss of precision
in intermediate calculations.