Added a script to compare multiple revisions.
The script calls the validation for the given git-revisions
after checking out the git hashes and calling scons.
Therefore it helps to quickly see the impact of a given change
to the framework.
It is optimized for tracking studies in the moment as it will
reuse (link) the EvtGen* files automatically.