Fix cppcheck error in LogPythonInterface
I thought I had included this in the restructuring pr but it seems not:
The `#ifdef #elif` was wrong. However I cannot recall why I did it this
weirdly so I changed the dispatch method to return void and return None
in the logging methods themselves. Probably I wanted to return
True/False if the message was sent but never did that so lets go for
simpler `#ifdef` structure.