Add framework/utilities/ScopeGuard.h with various factory functions to create RAII guards
I find myself reimplementing RAII guards in different places for
different use cases so I decided it would probably be a good idea to
collect some generic ones in a central place.

We already had one for the current working directory so I replaced that
one with a generic `ScopeGuard` class and added some static factory
functions

- `guardValue`: reset a object to its original value when the guard goes out of scope
- `guardGetterSetter`: similar to `guardValue` but for the case that its
  not a simple reference but needs a getter and setter function
- `guardFunctor`: similar to `guardGetterSetter` but for the case that
  getter and setter functions are both part of the same object as
  `operator()` overloads.
- `guardDeletion`: make sure a raw pointer is deleted when the guard goes
  out of scope. This feels redundant in lieu of `std::unique_ptr` but
  ROOT sometimes requires a T*&, will modify the pointer and allocate
  memory and then still give us ownership of the object. So
  `std::unique_ptr` will not work in these cases
- `guardStreamState`: Resets the flags, precision, width and fill
  character of a stream when the guard goes out of scope
- `guardWorkingDirectory`: Resets the working directory when the guard
  goes out of scope.

All these factory functions have the `[[nodiscard]]` attribute to create
warnings on misuse (the returned object needs to stay alive for the
guard to work) and depend on the fact that C++17 guarantees copy elision
for prvalues so returning temporary, unnamed objects doesn't need
copy/move construction or assignment operators.

The `ScopeGuard` class has a `release()` method to cancel cleanup. This
can be used for error handling: One might want to change some value and
check for error conditions and if there are no errors keep the new
value.

This guard very loosely follow the approach by the [p0052
draft](http://www.open-std.org/jtc1/sc22/wg21/docs/papers/2017/p0052r5.pdf)
to add `<scope>` to the C++ standard but the classes here are much much
simpler, mostly for maintainability.

We already have the `IOIntercept::InterceptorScopeGuard` for guarding IO
intercept but I don't see any use case nor should it be possible to
release this guard so it has not been modified.