Better implementation of pager.py
Don't use a temporary file but directly forward the file descriptors for
stdout and stderr to the pager process. This allows for a) incremental
updates and b) to capture output from C or C++

Now we can immediately use this feature in `b2conditionsdb dump` to show
the StreamerInfo in the payload to get detailed information about the
class structure stored in the payload (and its version).