A database class to hold the priors.
The prior pribability for one particle species is saved into a PIDPriorsTable
object, that consists of a 2-d std::vector containing the prior probability in bins of two  variables that the user preferes (typically p and cosTheta) , and std::vector for the errors on those priors, and two sorted 1-D std::vectors containing  the bin edges in the two variables.

The dbobject PIDPriors is nothing but a collection of six of these PIDPriorsTable objects (one per particle species), plus some wrappers for the getters and the setters.

Tests are provided for both the classes