Try to fix recurring problem with dictionaries not being rebuilt
While I haven't really found a reason for the failing rebuilt of
dictionaries two possible reasons couldd be a problem with the timestamp-md5 decider or
with our dictionary builder.

So to see if we can get rid of this problem we can switch back to the
default md5 provider but with a reduced max_drift of 10 minutes instead
of 2 days. (SCons will use cached md5sums if the file is older than 10
minutes).

Also I redid parts of the dictionary builder

* preserve original list of sources: If the RootDict builder was called
  with multiple sources these are now kept and additional headers are
  inserted before the linkdef. Allows to use the same builder for genfit
* add the rootmap and rdict.pcm files directly to the list of targets.
  This way SCons will know that one builder creates all three files and
  the dependencies should be more correct
* Use the new nodes returned for rootmap and pcm files for installation
* Use a local env variable in the builder call for the library name, no
  need to set it on the global environment
* rename cint -> cling

If we're lucky these changes might be enough to get rid of the problem
with failing incremental builds when data-/dbobjects are modified. If
not the changes don't seem to do any harm.