Changelog between release-02-01 and release-03 for the framework package
Also preparation for the changes from the analysis package so that we
can

* edit them independently
* can keep them up to date from now on (hopefully)

The idea is that

1. very important or general changes go directly in /whatsnew.rst
2. changes in the packages go into the txt files in the packages, the
librarians can make sure its filled poperly
3. one file per major release with "changes since" in master. That way in
light releases this should show up correctly.
4. for minor releases we might want an extra file but this should be
done on minor release branches I guess.
5. `.gitattributes merge=union` is set for the changelog .txt files.
This means there will be no conflicts in that file, git will use both
versions if there is a conflict. So two pull requests can add at the
same position in the changelog files.