Refactor `create_aliases_for_selected` (BII-4437)
This fixes a lot of issues with this function by completely rewriting
it and building up the full tree of selected particle paths before
trying to figure which particle paths are ambigious.

* it will be able to correctly figure out which part of the path is
  ambigious and only add indices for that part
* it will allow to also select the mother particle
* it does do some error checking and will raise ValueErrors if something
  is wrong (like invalid decay strings)

On top of that it now is much more customizable

* user can select between named and unnamed aliases
* users can include an index for **all** particles in named aliases
  (also for unambigious ones)
* users can select betweeen using the decay index or a simple relative
  indexing
* users can supply more than one custom prefix in one call

This also adds a set of  unit tests for this function.