Fix ranks if allowMultiRank=True
The check for previous values was not correct: For the first candidate
the previous value was set to the current value and then the rank was
onlonly increased if previous value != value which is never true

So we had at least 2 candidates with rank 1 every. single. time.

This commit is the minimal required fix.