Don't create Belle2FileCatalog.xml by default
This commit disables the default creation of Belle2FileCatalog.xml in
RootOutput. The catalog can always be created later using
`b2file-catalog-add`. This will also reduce the warning about
overwriting entries in the catalog (as there is no catalog)

It also tries to make the info message in RooInput message a bit more
clear and not try to convert LFN to PFN which would mostly just return
the LFN. So let's call it LFN and also show the real filename as it was
given to basf2.

The FileCatalog will now not try to create an empty file when it is
initialized since we don't want to create empty files just for looking
up an entry if the catalog is missing.

`b2file-metadata-add` will only update the file in the
file catalog if it was already registered. So to be sure to register the
file run `b2file-catalog-add` before or after `b2file-metadata-add`.

The default of `b2file-merge` is now to not register the output in a
file catalog. A new command line argument `--add-to-catalog` can be
supplied if this is needed. The old `--no-catalog` option does still
exist but will be ignored.

A sideeffect of these changes is that the LFN for files created by
RootOutput is would have been by default an empty string since the
absolute pathname was set when registering in the FileCatalog. This is
now moved to `b2file-merge` and RootOutput (which also makes sure this is
handled correctly in case the filename is a URL with a protocol
specifier).