Update sorting of datastoreprinter members
We did some weird sorting by the text representation of the member name
+ all arguments which works but sorts numbers as (0,1,10,2,...).

But python sort is stable so we can just sort by member name to get the
dictionary order (the only reason we sorted in the first place) out of
the way. Then the members are sorted by name and different argument
calls are kept as they are defined.