Started renaming/splitting the HLT trigger stages.
The final goal is to have only a single trigger stage (called "filter")
doing all the cuts needed for deciding whether to keep or dismiss the
event.
Any menu calculations used for skimming (called calib before) is done
afterwards and has no impact on the trigger result.

This changes the behavior of the trigger and simplifies it - althought
it is possible that we need to go back to a multi-stage approach later
with full luminosity.