framework: Add new RootFileInfo class to RootIOUtilities namespace
This class should be a common interface to work with files created by
RootOutput. It allows to conveniently check whether a file has the
persistent and the event tree and to obtain the metadata or the list of
branches in the tree.

It also allows to check whether a file contains all branches from a
given set to be able to check compatibility between files.