Disable usage of INVALID global tags and bail if the global tag doesn't exist
This is the "maximum effort" (=minimal code change needed) to do these
two things: After obtaining a list of payloads (which for missing global
tags is an empty list) we can actually run a second request to check the
status of the gt.

This second request is done after because in case of multiple servers we
depend on a 200 response to decide which is a working server.

We only do it once for a global tag an remember a static list of all
already checked global tags and we bail if the global tag info cannot be
obtained or if the status is INVALID.

We cannot enforce PUBLISHED without having a running tag and a
solution to allow testing states