Add path_utils args for setting CDC QI, disabling it by default
Now users can enable/disable the CDC track quality indicator from
`add_trackfinding_cdc` and also set the ID/path to the weightfile. If no
weightfile ID is chosen, the default is used, which I also changed, so that it
refers to a payload-like ID. However, at the moment of this comment, no payload
is in the conditions DB, so experimenting users/devs are expected to provide
their own weightfile.