Fix building on Ubuntu 18.04
After merging of pr #861 the software doesn't build on newer Ubuntu
versions anymore. The reason is a incorrect usage of the `wait()`
function to wait for a subprocess

You need to include `sys/wait.h` and call `wait(nullptr)`.

What the code is doing now is wrong. It actually does not wait for the
child at all but it just constructs a default constructed instance of
the `wait` union declared in `bits/waitstatus.h` which is then thrown
away and probably optimized away. So it doesn't do anything.

So alternatively you can just remove the wait call ...