Update VariablesToHistogram to also use the RootFileCreationManager
This allows to have ntuples and histograms and also multiple histograms
from different particle lists in the same file.

To make sure we don't have conflicts it allows to change the directory
of the histograms with a new parameter.

Also, try to correctly use the TDirectory::TContext scope guard to avoid
side effects by changing the global gDirectory in modules ...