PyObjConvUtils: Also support std::set for parameters (and other things)
Apparently this was missing but I wanted it. sue me.

This also now makes parameters slightly more flexible: `foo=4` will no
work if a list is expected as long as the one value is of the correct
type. Also for sets, `foo='string'` will be converted to a set with one
element if possible.