Add flag to toggle eventwise features in track QE MVA application
Before that, it was only possible in the collection module. Question: What does
happen if the MVA teacher does not train on the eventwise features, but they are
still added to the variable set in the MVA module? Then, they are not in the
weightfile. Are they just being ignored?