Geometry: Allow density scaling for systematic studies
This adds the backend changes required to rescale the density of all
Materials created with the `Belle2::geometry::Materials` interface.
There's also a new special module, `DensityScaler` which takes as
parameters the names of the geometry components and optionally a set of
material names to not scale and then should transparently affect only
the regions specified.

As catch-all the component '*' can be specified to scale everything. If
a component is explicitely specified together with '*' then the two
factors are multiplied

Warning: This will not work for some CDC materials which are defined
manually. And also the reweightable service geometry does not get
scaled.