GT Replay Part 1: Add new struct to contain PayloadMetadata
And also the class responsible to read/write them and get them from local text files for testing.

These two classes are so far not used but:

* the PayloadMetadata will be used everywhere to identify a payload. It replaces the previous "DBQuery" struct and also a few other structs which kept various parts of metadata information

* the TestingPayloadStorage is the replacement for the local database: Allow to create payloads and store the information in a plain human readable text file as well as provide payloads from these text files to be used in validation.

It removes a few features from the local database which were very seldomly (or ever) used:

* payload files must now be in the same directory as the text file
* there is no concept of 'default payload' which is tried if no valid payload is found for a given IoV

It is now much shorter then the previous implementation, mostly because inheritance and code duplication could be avoided. Reading of the payload file is now done lazy instead of construction time but missing files create errors as we will not have a "default" local database anymore (once all this is done)

It still depends on fnctl file locks to avoid race conditions when creating payloads which might not be the most robust choice. We might try to use `O_CREAT | O_EXCL` opening of the files to get another level of protection.