GT Replay, Part 3d: Add PayloadProvider to find payload files
This provider gets configured with a list of directories (or https urls)
and will in turn try to find the payload in each of them.

It wil always first look in the cache directory for downloaded payloads,
then go through all the user provided locations and finally try to
download from the server

For each location which is remote (custom http(s) mirror or central
server) it will try to download in the cache directory using file locks.
If that fails it will try once more to download it into a completely
random temporary filename. If that still fails we go on to the next
source.