GT Replay, Part 3a: Add Conditions Database Configuration object.
This commit adds a configuration object for all things conditions db. It
is supposed to live in `basf2.conditions` and provide all configuration
for the new system in one place.

This class is also to be used by the input modules to supply global tag
replay information to the database.

Most of the important documentatiopn is in the lower part of
Configuration.cc which contains all the python method descriptions.