GT Replay: Part 2a: Add base class for metadata providers
This is the base class for metadata providers: Classes responsible to
find payload information either from the central database server or
locally somehow

The planned providers are the central one and using a sqlite database
dump of certain global tags which will come in later stages ;)

The structure is as follows: The `MetadataProvider` always works on a
list of global tags. When the list is known it is given to the provider
to check if all tags exist and a valid for use. If not, processing will
be aborted.

If all globaltags are usable then the framework will ask the provider to
get the metadata information for a given list of payloads (via
`PayloadMetadata` instances which have only their name set) and the
provider will try to fill all the content for the given exp/run number.

Internally, it will automatically cache the metadata for a given run and
only ask the implementations if the cache is out of date.

Implementations only have to implement two methods

1. return the status of a given globaltag as string (or "" in case of
error)
2. update the list of payloads for a given global tag, exp number and
run number and call `addPayload` for each found payload.