New http(s) Downloader class for the conditions
This is copy and massive slim down of ConditionsPayloadsDownloader in the same
directory. Changes to the old one

* get rid of PayloadInfo struct -> now covered by PayloadMetadata
* don't use a custom scope guard for the curl session, we have a scope guard object now
* it's not really a singleton but there is a default instance which can be used to support session sharing across different parts (metadata, payload download)
* no longer handle payload metadata at all, just allow downloading of urls into std::ostream

In fact this is now a rather thin wrapper around libcurl to just be able to download a url into a stream and optionally verify the checksum of a stream. There are also two helper functions to escape strings to be safe to be used in urls and to join to strings with exactly one slash between them.

Everything else is now handled by metadata provider or payload providers.