Auto detect directory structure
For local payloads we now enforce a strict naming scheme and no longer
depend on the value of `payloadUrl` provided by the database.

- flat: same as before but guaranteed to be `dbstore_{name}_rev_{revision}.root`
- hashed: `{checksum[:2}/{name}_r{revision}.root`

We also try both, flat directory and hashed directory when looking for
local payloads. So the user doesn't have to configure anything and we
are still backwards compatible with flat directories.

This means we no longer use the existing `EConditionsDirectoryStructure`
enum but an updated inline one. The existing one will be deleted when
finalizing migration.