Add a test to mdst to make sure we don't accidentally modify classes or ClassDef versions
This test will make sure the StreamerInfo of mdst objects doesn't change
without us noticing and having to fix the expected values by hand.

This also gets rid of the "Do I have to increase ClassDef Version"
question.

In theory we should think about extending this check to all
classes for which we create a dictionary (CMS has something like this as
part of the build system) but that would be more long term