GT Replay, Part 5: enable global tag replay for all input modules
This is a rather simple commit which basically just makes all input
modules notify the database about the globaltags from the input file.

This also fixes BII-3862 as we now accumulate the number of MC events
correctly in the beginning and don't have to modify it later on.

It also makes sure that all files have the same branches (or at least
the same branches as the first file) so we don't run into weird crashes
if file number 5 is missing one branch. There was a question about this
problem once but I cannot find an issue.